🚚 chore: rename directory name under mydashboard to avoid routing
