+++ src/app/(with-header-sidebar)/mydashboard/_types/dashboards.ts
