+++ src/components/header/UserInfo.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import type { User } from '@/app/(with-header-sidebar)/mydashboard/types/user';
+import type { User } from '@/app/(with-header-sidebar)/mydashboard/_types/user';
 import Image from 'next/image';
-import useWindowSize from '@/app/(with-header-sidebar)/mydashboard/hooks/useWindowSize';
+import useWindowSize from '@/app/(with-header-sidebar)/mydashboard/_hooks/useWindowSize';
 import { useState, useEffect } from 'react';
 import UserInfoSkeleton from '../skeleton/UserInfoSkeleton';
 import styles from './UserInfo.module.css';
